+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Changes to the plugin.
 
 - Missing choices no longer make document symbols vanish.
 - Variables whose name includes `param_count` (but are _not_ `param_count` itself) are now tracked properly.
+- `*text_image` is now parsed properly.
 
 ## [2.6] - 2022-04-16